# caddyfile: Switch to slices.Equal for better performance