@francislavoie 2024-01-25
Interesting :sweat_smile:

> Or we can simply use slices.Equal from "golang.org/x/exp/slices".

That would be fine too, if you prefer.