@bbaa-bbaa 2024-01-25
> That would be fine too, if you prefer.
 Switch to slices.Equal.